+++ admin/src/constants/settings.ts
@@ -1,6 +1,8 @@
 import { type WebsiteSettings } from '../types/settings'
 
 export const DEFAULT_WEBSITE_SETTINGS: WebsiteSettings = {
+  security: { recaptchaEnabled: false, recaptchaSiteKey: '', recaptchaSecretKey: '' },
+  email: { enabled: false, smtpHost: '', smtpPort: 587, smtpSecurity: 'starttls', smtpUsername: '', smtpPassword: '', fromName: 'Surf Mangalore', fromEmail: '', adminNotificationEmail: '', replyToEmail: '' },
   about: {} as WebsiteSettings['about'],
   experiencePage: {} as WebsiteSettings['experiencePage'],
   lessonPage: {} as WebsiteSettings['lessonPage'],